NF: moving waitForAllToFinish to TaskManager
