+++ AnkiDroid/src/main/java/com/ichi2/async/TaskManager.java
@@ -3,6 +3,8 @@ package com.ichi2.async;
 import android.content.res.Resources;
 import android.os.AsyncTask;
 
+import com.ichi2.utils.ThreadUtil;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.LinkedList;
@@ -138,6 +140,29 @@ public class TaskManager {
 
     public static ProgressCallback progressCallback(CollectionTask task, Resources res) {
         return new ProgressCallback(task, res);
+    }
+
+
+    /**
+     * Block the current thread until all CollectionTasks have finished.
+     * @param timeoutSeconds timeout in seconds
+     * @return whether all tasks exited successfully
+     */
+    @SuppressWarnings("UnusedReturnValue")
+    public static boolean waitForAllToFinish(Integer timeoutSeconds) {
+        // HACK: This should be better - there is currently a race condition in sLatestInstance, and no means to obtain this information.
+        // This should work in all reasonable cases given how few tasks we have concurrently blocking.
+        boolean result;
+        result = waitToFinish(timeoutSeconds / 4);
+        ThreadUtil.sleep(10);
+        result &= waitToFinish(timeoutSeconds / 4);
+        ThreadUtil.sleep(10);
+        result &= waitToFinish(timeoutSeconds / 4);
+        ThreadUtil.sleep(10);
+        result &= waitToFinish(timeoutSeconds / 4);
+        ThreadUtil.sleep(10);
+        Timber.i("Waited for all tasks to finish");
+        return result;
     }
